requirements.txt
api==0.0.6
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
curve==0.1.0
cycler==0.10.0
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
line-bot-sdk==1.5.0
MarkupSafe==1.0
matplotlib==2.1.0
nose==1.3.7
numpy==1.13.3
pandas==0.20.3
peewee==2.10.2
psycopg2==2.7.3.1
pyasn1==0.3.4
PyMySQL==0.7.11
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
rsa==3.4.2
six==1.10.0
thrift==0.10.0
urllib3==1.22
Werkzeug==0.12.2
beautifulsoup4==4.6.0

bs4==0.0.1
PyLyrics==1.1.0